% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subir_bq.R
\name{subir_bq}
\alias{subir_bq}
\title{Função para subir base no Goosle Big Query}
\usage{
subir_bq(da, nome)
}
\arguments{
\item{da}{base.}

\item{nome}{nome desejado.}
}
\description{
Essa função conecta com o google big query e escreve a
data base com um nome específico
}
