% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{tjsp2inst}
\alias{tjsp2inst}
\title{tjsp2inst}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 612517 rows and 14 columns.
}
\usage{
tjsp2inst
}
\description{
tjsp2inst
}
\keyword{datasets}
